# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ScreenTinker Public API
  version: 2.3.2
  description: |
    Public, token-scoped REST API for ScreenTinker digital signage.

    ## Authentication
    Every endpoint (except the two unauthenticated device-render endpoints) accepts a
    **scoped personal access token** presented as `Authorization: Bearer st_...`. The
    dashboard's session JWT is also accepted, but this contract targets token consumers.

    ## Scope ladder (read &#60; write &#60; full)
    Tokens carry one of three scopes; higher scopes subsume lower ones:
    - `read`  — all `GET` endpoints.
    - `write` — all resource mutations (`POST` / `PUT` / `PATCH` / `DELETE`).
    - `full`  — operational fleet commands (currently only
      `POST /groups/{id}/command`: reboot / shutdown / screen on/off).

    Each operation advertises its minimum scope via the `x-required-scope` extension
    and restates it in the operation description. A token with insufficient scope is
    rejected with `403`.

    All resources are scoped to the token's workspace; cross-workspace access is not
    available to tokens.
servers:
  - url: /api
...
security:
  - ApiToken: []
  - SessionJWT: []
components:
  securitySchemes:
    ApiToken:
      type: http
      scheme: bearer
      bearerFormat: "st_<token>"
      description: |
        Scoped personal access token. Send as `Authorization: Bearer st_...`.
        The token's scope (`read`, `write`, or `full`) gates which operations
        it may call (see the scope ladder in the API description).
    SessionJWT:
      type: http
      scheme: bearer
      description: |
        Dashboard session JWT. Also accepted on every authenticated endpoint,
        but this spec targets scoped-token consumers.
  schemas:
...
    DeckDocument:
      type: object
      description: |
        The editable source for a deck. Each slide carries a **template** (the view: element
        geometry, style and motion, each with a `slot` name) and **fields** (the record: `{slot:
        value}`). They are joined when the slide is rendered, which is what makes it possible to
        come back and change a headline without rebuilding the layout.
      properties:
        slides:
          type: array
          maxItems: 100
          items:
            type: object
            properties:
              id:
                type: string
                description: 'Unique within the deck — it is how a republish finds the widget it already made.'
              name: { type: string, maxLength: 120 }
              dwell_sec:
                type: integer
                minimum: 1
                maximum: 3600
                default: 10
                description: |
                  How long the slide stays up. A slide's motion has to finish inside this; if it
                  does not, save and publish both return a `motion-outlives-dwell` warning naming
                  the slide. Warned rather than refused.
              widget_id:
                type: [string, 'null']
                description: 'Assigned by publish. Present so a republish updates rather than duplicates.'
              template:
                type: object
                properties:
                  background: { type: string, description: 'Hex colour. Anything else falls back to black.' }
                  elements:
                    type: array
                    maxItems: 40
                    items:
                      type: object
                      properties:
                        slot: { type: string, description: 'The key this element reads from fields.' }
                        kind: { type: string, enum: [head, body, stat, image, rule, box] }
                        box:
                          type: object
                          properties:
                            x: { type: number }
                            y: { type: number }
                            w: { type: number }
                            h: { type: number }
                        content_id: { type: [string, 'null'], description: 'For an image — a content row in this workspace.' }
                        style:
                          type: object
                          properties:
                            color: { type: string, description: 'Hex. Anything else falls back to white.' }
                            font:
                              type: string
                              description: |
                                A bundled family (`inter`, `archivo`, `oswald`, `bitter`,
                                `jetbrains-mono`), an uploaded one (`u:<uuid>`), or one of the
                                legacy generics (`sans`, `serif`, `mono`, `condensed`) which still
                                resolve.
                            size_cqw:
                              type: number
                              description: 'Relative to the screen, never pixels — so a slide scales from 720p to 4K.'
                            weight: { type: integer, minimum: 100, maximum: 900 }
                            align: { type: string, enum: [left, center, right] }
                            opacity: { type: number, minimum: 0, maximum: 1 }
                            radius_cqw: { type: number }
                        motion:
                          type: [object, 'null']
                          description: 'Null means the element is simply present from the first frame.'
                          properties:
                            animation: { type: string, enum: [fade, slideL, slideR, slideU, slideD, zoom, wipe] }
                            delay: { type: number, minimum: 0, maximum: 30 }
                            duration: { type: number, minimum: 0.05, maximum: 10 }
                            easing: { type: string, enum: [ease-out, ease-in, ease-in-out, linear, soft] }
              fields:
                type: object
                description: 'The values, keyed by slot. Editing one of these does not touch the template.'
                additionalProperties: { type: string }
...
paths:
...
  /slide-decks:
    get:
      tags: [slides]
      summary: List the slide decks in this workspace
      description: |
        A deck is the editable source document for a set of slides. Publishing one emits a playlist
        of slide widgets — so decks take no part in playback, and everything downstream
        (scheduling, groups, device assignment) works on the playlist, not on this.
      x-required-scope: read
      responses:
        '200': { description: 'An array of decks with slide counts and total duration.' }
    post:
      tags: [slides]
      summary: Create a deck
      x-required-scope: write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 120 }
                doc: { $ref: '#/components/schemas/DeckDocument' }
      responses:
        '201': { description: 'The created deck, with any publish warnings.' }
        '400': { description: 'A deck needs a name.' }
  /slide-decks/{id}:
    get:
      tags: [slides]
      summary: Fetch a deck document
      x-required-scope: read
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: 'The deck, its document, and its warnings.' }
        '404': { description: 'No such deck in this workspace.' }
    put:
      tags: [slides]
      summary: Save a deck document
      description: |
        Saving is **not** publishing — nothing reaches a screen until you publish. Warnings come
        back on every save so an editor can show a problem while somebody works, rather than
        refusing the change.
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 120 }
                doc: { $ref: '#/components/schemas/DeckDocument' }
      responses:
        '200': { description: 'The saved deck.' }
        '404': { description: 'No such deck in this workspace.' }
    delete:
      tags: [slides]
      summary: Delete a deck document
      description: |
        **The published playlist and its widgets are kept.** By the time a deck has been published
        what is on the wall is a playlist, possibly assigned to devices or on a schedule — deleting
        the document must not take those down with it. Pass `with_published=1` to remove them too,
        which still refuses to delete a widget another playlist is playing.
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - name: with_published
          in: query
          required: false
          schema: { type: string, enum: ['1', 'true'] }
          description: 'Also remove the slide widgets this deck published.'
      responses:
        '200': { description: '{ success, removed_widgets, playlist_kept }.' }
        '404': { description: 'No such deck in this workspace.' }
  /slide-decks/{id}/publish:
    post:
      tags: [slides]
      summary: Publish a deck to a playlist of slide widgets
      description: |
        Creates or updates one slide widget per page and a playlist that orders them, then publishes
        that playlist. Republishing updates the same widgets rather than making new ones, and a
        slide removed from the deck has its widget deleted — unless another playlist still plays it,
        in which case it is only unlinked.
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: 'The deck, plus published.playlist_id, published.slides and published.removed.' }
        '404': { description: 'No such deck in this workspace.' }
        '500': { description: 'The deck could not be published.' }

  # -------------------------------------------------------------- slide fonts
...